+++ r-package/tests/testthat/test-setup_r5.R
@@ -1,6 +1,6 @@
 context("setup_r5")
 
-testthat::skip_on_cran()
+# testthat::skip_on_cran()
 
 path <- system.file("extdata/poa", package = "r5r")
 
@@ -9,6 +9,8 @@ test_that("setup_r5 - expected behavior", {
 
   testthat::expect_message( setup_r5(data_path = path, verbose = F) )
   testthat::expect_message( setup_r5(data_path = path, verbose = T) )
+
+  testthat::expect_message( setup_r5(data_path = path, use_elevation=T) )
 
   # remove files GTFS
   #  file.rename(file.path(path, "poa.zip"), file.path(path, "poa.x"))
@@ -28,8 +30,9 @@ test_that("setup_r5 - expected errors", {
   testthat::expect_error( setup_r5(data_path = NULL) )
   testthat::expect_error( setup_r5(data_path = 'a') )
   testthat::expect_error(setup_r5(data_path = path, verbose = 'a'))
+  testthat::expect_error(setup_r5(data_path = path, temp_dir = 'a'))
   testthat::expect_error(setup_r5(data_path = path, use_elevation = 'a'))
-  testthat::expect_error(setup_r5(data_path = path, version = 'a'))
+#  testthat::expect_error(setup_r5(data_path = path, version = 'a'))
 
   # No OSM data
   testthat::expect_error( setup_r5(data_path = file.path(.libPaths()[1]) ) )
@@ -41,5 +44,23 @@ test_that("setup_r5 - expected errors", {
   #   file.rename(file.path(path, "network2.x"), file.path(path, "network.dat"))
 
   })
+
+test_that("'overwrite' parameter works correctly", {
+
+  testthat::expect_error(setup_r5(path, overwrite = 1))
+
+  # since a network was already created, if overwrite = FALSE it should use it
+  testthat::expect_message(
+    r5r_core <- setup_r5(path, verbose = FALSE),
+    regexp = "Using cached network\\.dat from "
+  )
+
+  # but if overwrite = TRUE, then it should create a new network anyway
+  testthat::expect_message(
+    r5r_core <- setup_r5(path, verbose = FALSE, overwrite = TRUE),
+    regexp = "Finished building network\\.dat at "
+  )
+
+})
 
 stop_r5()